+++ Changelog.md
@@ -1,3 +1,16 @@
+# 1.2.0 #
+
+## Added ##
+
+* Added new method to the `Quorrax\Classes\Variable` class:
+  * Added the `isFloat` public method.
+  * Added the `isInteger` public method.
+  * Added the `isString` public method.
+* Added new method to the `Quorrax\Interfaces\Variable` interface:
+  * Added the `isFloat` public method.
+  * Added the `isInteger` public method.
+  * Added the `isString` public method.
+
 # 1.1.0 #
 
 ## Added ##
@@ -10,10 +23,10 @@
 ## Added ##
 
 * Added the `Quorrax\Classes\Variable` class.
-    * Added the `setValue` private method.
-    * Added the `variable` private property.
-    * Added the `getValue` public method.
-    * Added the `__construct` public method.
+  * Added the `setValue` private method.
+  * Added the `variable` private property.
+  * Added the `getValue` public method.
+  * Added the `__construct` public method.
 * Added the `Quorrax\Interfaces\Variable` interface.
-    * Added the `getValue` public method.
-    * Added the `__construct` public method.
+  * Added the `getValue` public method.
+  * Added the `__construct` public method.